+++ src/app/(providers)/stray-dogs/_component/page.module.scss
@@ -1,0 +1,188 @@
+.filterWrap {
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin: 30px auto 50px;
+}
+
+.filterContent {
+  flex-direction: column;
+  .region {
+    display: flex;
+    align-items: center;
+    gap: 15px;
+
+    .regionWrapper {
+      display: flex;
+      gap: 15px;
+    }
+  }
+  p {
+    font-weight: 700;
+    font-size: 32px;
+    margin-bottom: 20px;
+  }
+}
+
+.searchButton {
+  width: 45px;
+  height: 45px;
+  background-color: #f7f9f9;
+  border: 1px solid #979797;
+  border-radius: 5px;
+  color: black;
+}
+
+.calender {
+  width: 100%;
+  margin: auto;
+  display: flex;
+  gap: 15px;
+}
+
+.datePicker {
+  width: 200px;
+  height: 45px;
+  border: 1px solid #979797;
+  border-radius: 5px;
+  padding-left: 5px;
+  text-align: center;
+  cursor: pointer;
+}
+
+.selectCity {
+  width: 200px;
+  height: 45px;
+  border: 1px solid #979797;
+  border-radius: 5px;
+  padding: 10px;
+  appearance: none;
+}
+
+.gridContainer {
+  padding: 2rem 0;
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  gap: 2rem;
+  margin-top: 30px;
+}
+
+.listCard {
+  width: 100%;
+  cursor: pointer;
+}
+
+.imageWrap {
+  margin: 0 auto 0 auto;
+  overflow: hidden;
+  cursor: pointer;
+  border-radius: 13px;
+  aspect-ratio: 1 / 1;
+}
+.imageWrap::after {
+  display: block;
+  padding-bottom: 100%;
+}
+
+.image {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.imageWrap:hover {
+  border-radius: 13px;
+  box-shadow: 0px 0px 10px 0px rgba(0, 0, 0, 0.4);
+}
+
+.explanationWrap {
+  display: flex;
+  width: 250px;
+  margin: 10px auto;
+}
+
+.titleColumn {
+  display: flex;
+  justify-content: flex-start;
+  flex-direction: column;
+  align-items: flex-start;
+  height: 200px;
+  width: 120px;
+  margin: auto 0;
+  p {
+    font-size: 15px;
+    font-weight: 700;
+    line-height: 30px;
+  }
+}
+
+.contentColumn {
+  display: flex;
+  justify-content: flex-start;
+  flex-direction: column;
+  align-items: flex-start;
+  height: 200px;
+  margin: auto 0 auto 15px;
+  width: 100%;
+  p {
+    font-weight: 500;
+    font-size: 15px;
+    line-height: 30px;
+  }
+}
+
+@media screen and (max-width: 1200px) {
+  .filterWrap {
+    flex-direction: column;
+    align-items: flex-start;
+    justify-content: space-evenly;
+    width: 100%;
+    height: 250px;
+    gap: 40px;
+    padding: 0 2rem;
+    margin: 0 30px 50px;
+  }
+  .gridContainer {
+    padding: 0 2rem;
+    grid-template-columns: repeat(3, 1fr);
+  }
+}
+@media screen and (max-width: 820px) {
+  .filterWrap {
+    height: 250px;
+  }
+  .gridContainer {
+    grid-template-columns: repeat(2, 1fr);
+  }
+}
+
+@media screen and (max-width: 480px) {
+  .filterWrap {
+    margin: 0 auto 30px auto;
+    height: 250px;
+  }
+  .gridContainer {
+    grid-template-columns: repeat(1, 1fr);
+  }
+  .filterContent {
+    p {
+      font-size: 25px;
+      margin-bottom: 15px;
+    }
+  }
+  .datePicker {
+    width: 160px;
+    height: 45px;
+  }
+  .selectCity {
+    width: 160px;
+    height: 45px;
+  }
+  .region {
+    flex-direction: column;
+    .searchButton {
+      width: 100%;
+    }
+  }
+}
